+++ src/js/graphcreator.js
@@ -63,7 +63,7 @@ window.Widget = (function () {
         container.appendChild(group_title);
 
         // Create the group column select
-        this.group_axis_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.group_axis_select = new StyledElements.Select({'class': 'full'});
         this.group_axis_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.group_axis_select);
 
@@ -84,35 +84,35 @@ window.Widget = (function () {
         group_title = document.createElement('h3');
         group_title.innerHTML = 'Axis X';
         container.appendChild(group_title);
-        this.axisx_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.axisx_select = new StyledElements.Select({'class': 'full'});
         this.axisx_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.axisx_select);
 
         group_title = document.createElement('h3');
         group_title.innerHTML = 'Axis Y';
         container.appendChild(group_title);
-        this.axisy_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.axisy_select = new StyledElements.Select({'class': 'full'});
         this.axisy_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.axisy_select);
 
         group_title = document.createElement('h3');
         group_title.innerHTML = 'Size Axis';
         container.appendChild(group_title);
-        this.axisz_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.axisz_select = new StyledElements.Select({'class': 'full'});
         this.axisz_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.axisz_select);
 
         group_title = document.createElement('h3');
         group_title.innerHTML = 'Series field';
         container.appendChild(group_title);
-        this.series_field_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.series_field_select = new StyledElements.Select({'class': 'full'});
         this.series_field_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.series_field_select);
 
         group_title = document.createElement('h3');
         group_title.innerHTML = 'Name of the bubble';
         container.appendChild(group_title);
-        this.id_bubble_select = new StyledElements.StyledSelect({'class': 'full'});
+        this.id_bubble_select = new StyledElements.Select({'class': 'full'});
         this.id_bubble_select.addEventListener('change', create_graph_config.bind(this));
         container.appendChild(this.id_bubble_select);
     };
@@ -216,10 +216,8 @@ window.Widget = (function () {
         preferences.graph_series = JSON.stringify(preferences.graph_series);
         preferences.graph_fields = JSON.stringify(preferences.graph_fields);
 
-        var name = document.getElementById("dashboard_name").value;
-
         MashupPlatform.mashup.createWorkspace({
-            name: name,
+            name: this.nameinput.getValue(),
             mashup: 'CoNWeT/' + mashupname + '/1.0.0',
             preferences: preferences,
             onSuccess: function (workspace) {
@@ -235,7 +233,7 @@ window.Widget = (function () {
     };
 
     Widget.prototype.init = function init() {
-        this.layout = new StyledElements.StyledNotebook({'class': 'se-notebook-crumbs'});
+        this.layout = new StyledElements.Notebook({'class': 'se-notebook-crumbs'});
         this.layout.insertInto(document.body);
 
         var chart_tab = this.layout.createTab({name: "1. Chart", closable: false});
@@ -248,24 +246,18 @@ window.Widget = (function () {
         var mashup_panel = document.createElement('div');
         mashup_panel.className = "mashup_panel";
         var temp_p = document.createElement('p');
-        temp_p.style = "margin-bottom: 10px";
         temp_p.appendChild(document.createTextNode("If you have finished, you can create a new dashboard now."));
         mashup_panel.appendChild(temp_p);
 
-        var temp_d = document.createElement("div");
+        var temp_d = document.createElement("label");
         temp_d.appendChild(document.createTextNode("Name:  "));
 
-        var name_input = document.createElement("input");
-        name_input.type = "text";
-        name_input.value = "CKAN Wirecloud View";
-        name_input.className = "se-text-field";
-        name_input.id = "dashboard_name";
-
-        temp_d.appendChild(name_input);
+        this.nameinput = new StyledElements.TextField();
+        this.nameinput.setValue("CKAN Wirecloud View").appendTo(temp_d);
 
         mashup_panel.appendChild(temp_d);
 
-        this.googleButton = new StyledElements.StyledButton({text: 'Create dashboard with Google Graph', class: 'btn-primary'});
+        this.googleButton = new StyledElements.Button({text: 'Create dashboard with Google Graph', class: 'btn-success'});
         this.googleButton.insertInto(mashup_panel);
 
         this.workspace_tab.appendChild(mashup_panel);
@@ -276,7 +268,7 @@ window.Widget = (function () {
 
         this.googleButton.disable();
 
-        this.flotr2Button = new StyledElements.StyledButton({text: 'Create dashboard with flotr Graph', class: 'btn-primary'});
+        this.flotr2Button = new StyledElements.Button({text: 'Create dashboard with flotr Graph', class: 'btn-success'});
         this.flotr2Button.insertInto(mashup_panel);
 
         this.workspace_tab.appendChild(mashup_panel);
@@ -289,7 +281,7 @@ window.Widget = (function () {
 
         // create the data selection tab
         // The graph selection tab depends on the data selection tab, so it must be build first
-        this.data_form_alternatives = new StyledElements.StyledAlternatives();
+        this.data_form_alternatives = new StyledElements.Alternatives();
         this.data_tab.appendChild(this.data_form_alternatives);
 
         this.normal_form_alternative = this.data_form_alternatives.createAlternative();